feat(sector_map): always show hyperspace target disk
Currently the hyperspace target disk is hidden if the system is the
currently selected system.

This change increases the size of the hyperspace target disk and always
shows it. The disks are now shown, from top-to-bottom, as the currently
selected system (green), the current hyperspace target system (grey), and
the current system (blue).