feat(radar): add manual zoom mode to scanner
The radar is zoomed manually, but the scanner has automatic zoom which
can lead to some issues while trying to get an overview of the current
area around the ship as it is usually zoomed out too far.

This commit adds key bindings to switch between the radar and the
scanner, as well as bindings to zoom the scanner and reset the mode back
to automatic.

The zoom bar of the scanner is moved up to make room for the
new manual/automatic zoom mode indicator on the left as well as the
current range indicator which is in the same position on the right as the
radar display.

The alternative is to display the range in the center above the zoom bar,
but then it is in a different position compared to the radar.

TODO:
* use icons for the automatic/manual mode indicator
* add hover-mode for the scanner to show ship information
* use keybindings instead of hardcoded keys