feat(radar): draw radar in UI colors
The radar ("azimuthal mode") is drawn using the reticule colours (white),
which clashes with the rest of the dashboard icons/controls as well as
the scanner ("3d mode").

This commit changes the radar colours to (mostly) match the scanner
colours.

It also tweaks the line-width of the scanner and radar backgrounds:
- rader inner lines are made thinner
- scanner border is made thicker