feat(radar): major cleanup refactor
This commit is a major refactor of the radar module code.

The scanner and radar were each turned into a class (table) to make the
main draw function more generic.
TODO: there's probably a good way to refactor this further to have
both instruments derive from a base class in order to ensure the same
base functionality.

Additionally:

* The 'onChangeMFD' event was removed

* The scanner (3D) and radar (2D azimuth) now each have their own zoom
  level. This prevents the radar from ending up with an auto-zoom zoom
  level when switching from the scanner since it only ever has manual
  zoom.

* The autopilot controls were also moved to the right a little bit to make
  more room for the zoom-level display.

* The scanner was made marginally taller.