feat(radar): add manual/automatic icons for radar mode
Add two new icons to icons.svg on the 21st row (starting on index 320)
for the radar's automatic and manual zoom mode indicator.

This necessitated modifying several other files to increase the grid size
of icons.svg in order for it to load properly. This is not an ideal
situation..

Co-authored-by: Szlrd <bszlrd@gmail.com>