feat(radar): !MESS! fix button and render using new icons
THIS COMMIT IS A RIGHT ROYAL MESS AND THE CODE NEEDS CLEANING UP. DO NOT
MERGE!!!

In order to get the new buttons near the radar to actually work, they
had to be rendered inside a window. They now display tooltips and
respond to mouse clicks.

HOWEVER, placing the entire randar into the window does not work as the
3D radar's background is not rendered when it's in the window (possibly
due to coordinate issues? Alternatively because it's rendered in C++
instead of LUA?) although its "pips" are.

The 2D radar is rendered properly even when inside a window.

Finally, the range indication text has to be rendered OUTSIDE the window
as otherwise it doesn't show up. It briefly did, but then it disappeared
again.

Lastly, the buttons now use the new icons added by Szlrd.