feat: move hyperspace origin into hyperspace cloud
The implementation uses the ship object's hyperspace destination for
both the destination as well as the origin, depending on whether the
ship was currently travelling to another system, or arriving in a
system.

Firstly this sort of semantic overload of a variable may have been
required in very resource constrained environments in the past but
has no place in modern development.

Secondly there are some code paths which seem to rely on the variable
actually containing the destination in order for the ship to behave
correctly in the destination system.

We fix this by ensuring a ship's hyperspace destination is never
overwritten with the origin system. Instead, the origin is saved into
the hyperspace cloud. This also makes the code clearer and more
self-documenting in places which use the hyperspace origin.